more stuff
- 'Stamina Attributes' is no longer a required dependency
- attack stamina_cost can now be a fractional value
- added global multipliers to attack stamina costs, one for attacks that are feinted, one for attacks that are actually executed. These multipliers can be set in the server config.
- added server config option to enable feinting to skip the current attack (Thank you, @ji.111 for the idea)
- fixed crash when attack damage_type field is not a valid damage type identifier, fixes #2
- clean up

+++ src/main/java/com/github/theredbrain/bettercombatextension/bettercombat/DuckWeaponAttributesAttackMixin.java
@@ -1,9 +1,9 @@
 package com.github.theredbrain.bettercombatextension.bettercombat;
 
 public interface DuckWeaponAttributesAttackMixin {
-    int bettercombatextension$getStaminaCost();
+    float bettercombatextension$getStaminaCost();
 
-    void bettercombatextension$setStaminaCost(int staminaCost);
+    void bettercombatextension$setStaminaCost(float staminaCost);
 
     String bettercombatextension$getDamageType();
